Image Captioning/requirements.txt
transformers
torch
gradio
Pillow
networkx==3.4
spotipy
google-api-python-client
